refactor: extract plugin logic into FilamentAccessManagementPlugin
